requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
flask-restx==1.3.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
numpy==1.26.3
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.1
rpds-py==0.17.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
Werkzeug==3.0.1
zipp==3.17.0
